fix: csr -> ssr
